python/projects/video-subtitiles/backend/transcription_translation/requirements.txt
google-cloud-storage==2.7.0
google-cloud-pubsub==2.13.0
speechrecognition==3.10.0
googletrans==3.1.0a0